backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
google-genai>=1.0.0
sqlalchemy>=2.0.36
pandas>=2.2.0
python-dotenv>=1.0.1
pydantic>=2.10.0
python-multipart>=0.0.9
openpyxl>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
mangum>=0.17.0
bcrypt>=4.0.0

